refactor: remove remaining core compatibility symlinks
