+++ project-structure-refactor-plan.md
@@ -28,7 +28,7 @@ This plan is based on the current repo layout and build system behavior (`build.
 | D1 | `boot/` to `core/boot/` | ✅ Completed | D1a + D1b completed: sources/common/include/discovery/protocols are canonical in `core/boot/*`; D1c completed: kernel sub-target CMake include wiring now uses migration-aware `BHARAT_BOOT_INCLUDE_DIR`; D1e completed: legacy `boot/{src,common,include,discovery,protocols}` symlink fanout removed after migrating in-repo build/test include paths to `core/boot/*`; D1f completed: kernel build wiring now requires canonical `core/boot/*` roots (no legacy boot root fallback in `core/kernel` CMake). |
 | D2 | `kernel/` to `core/kernel/` | ✅ Completed | D2a–D2f completed. Canonical kernel build now requires `core/kernel/*`; legacy `kernel/{CMakeLists.txt,include,src,staging,linker*.ld}` compatibility symlinks have been removed after canonical caller migration. |
 | D3 | `arch/` + `hal/` + `platform/` to `core/*` | ✅ Completed | D3a–D3c completed. Legacy top-level `arch`, `hal`, and `platform` compatibility symlinks have been removed; canonical roots are `core/arch`, `core/hal`, and `core/platform`. |
-| D4 | `lib/` + `stacks/` + `drivers/` + `services/` + `personalities/` to `core/*` (bounded slice) | ✅ Completed | D4a–D4b completed. Canonical roots remain `core/*`; top-level compatibility symlinks for `lib`, `stacks`, `drivers`, `services`, and `personalities` are still retained for transition-safe host test compatibility. |
+| D4 | `lib/` + `stacks/` + `drivers/` + `services/` + `personalities/` to `core/*` (bounded slice) | ✅ Completed | D4a–D4c completed. Canonical roots are now enforced under `core/*`; top-level compatibility symlinks for `lib`, `stacks`, `drivers`, `services`, and `personalities` have been removed after migrating in-repo callers/tests to canonical `core/*` paths. |
 | E1 | `tests/` to `quality/tests/` (bounded Phase 5 slice) | ✅ Completed | Canonical host/unit/e2e test tree now lives under `quality/tests`; root `tests` compatibility symlink has been removed after in-repo path migration. |
 | F1 | Public umbrella `include/` to `interface/include/` | ✅ Completed | Canonical public header tree now resides in `interface/include`; root `include` compatibility symlink has been removed after migrating in-repo CMake include wiring to the canonical interface path. |
 | F2 | `user/` apps+SDK to `experience/user/` | ✅ Completed | Canonical user-space tree now resides in `experience/user`; root `user` compatibility symlink has been removed and top-level CMake still prefers canonical path with fallback warning support. |
@@ -532,6 +532,7 @@ This section converts the phase model into concrete, code-aware slices from the 
   - D4a (completed): moved `lib/` to `core/lib/` and `stacks/` to `core/stacks/`; retained `lib` and `stacks` as compatibility symlinks.
   - Top-level build wiring now prefers canonical `add_subdirectory(core/lib)` and `add_subdirectory(core/stacks)` while legacy path references still resolve through symlinks.
   - D4b (completed): moved `platform`, then `drivers`, `services`, and `personalities` to canonical `core/*` roots with compatibility symlinks retained during transition.
+  - D4c (completed): migrated in-repo quality/experience callers off legacy top-level `lib|stacks|drivers|services|personalities` paths and removed the remaining root compatibility symlinks to enforce canonical `core/*` roots.
 
 ---
