+++ quality/tests/host/test_tlb_stress.c
@@ -11,8 +11,8 @@
 #include "../../kernel/include/bharat/cpu_local.h"
 #include "../../kernel/include/hal/hal.h"
 #include "../../kernel/include/urpc/urpc_bootstrap.h"
-#include "../../services/core/subsysmgr/include/bharat/msg/transport.h"
-#include "../../services/monitor/generated/bharat_monitor_v1_types.h"
+#include "../../core/services/core/subsysmgr/include/bharat/msg/transport.h"
+#include "../../core/services/monitor/generated/bharat_monitor_v1_types.h"
 
 // --- Mocking ---
 hal_tlb_ops_t *active_hal_tlb;
