+++ quality/tests/test_automotive_subsys.c
@@ -2,8 +2,8 @@
 #include <stdint.h>
 #include <stdio.h>
 
-#include "../drivers/include/bharat/drivers/can.h"
-#include "../services/core/subsysmgr/include/bharat/automotive/automotive.h"
+#include "../core/drivers/include/bharat/drivers/can.h"
+#include "../core/services/core/subsysmgr/include/bharat/automotive/automotive.h"
 
 static int g_started = 0;
 static int g_power_mode_events = 0;
