+++ quality/tests/test_integration_core_subsys.c
@@ -9,8 +9,8 @@
 #include "../kernel/include/sched/algo_matrix.h"
 #include "../kernel/include/subsystem_profile.h"
 
-#include "../personalities/linux/linux_compat.h"
-#include "../services/core/subsysmgr/include/bharat/automotive/automotive.h"
+#include "../core/personalities/linux/linux_compat.h"
+#include "../core/services/core/subsysmgr/include/bharat/automotive/automotive.h"
 
 #include "../include/bharat/ui/fb_render.h"
