+++ quality/tests/test_profile_edge.c
@@ -9,7 +9,7 @@
 #include "../kernel/include/sched/algo_matrix.h"
 #include "../kernel/include/subsystem_profile.h"
 
-#include "../services/core/subsysmgr/include/bharat/automotive/automotive.h"
+#include "../core/services/core/subsysmgr/include/bharat/automotive/automotive.h"
 
 #include "../include/bharat/ui/fb_render.h"
