+++ quality/tests/test_windows_compat_init.c
@@ -2,7 +2,7 @@
 #include <stdint.h>
 #include <stdio.h>
 
-#include "../personalities/windows/win_compat.h"
+#include "../core/personalities/windows/win_compat.h"
 
 int main(void) {
     printf("Running Windows Subsystem integration tests...\n");
